Fix build: add 'originator' param to Session.login_with_password
We made a change to the Session.login_with_password API call, so
that it takes an optional argument for the "originator" of the
API call. This is an optional argument on the wire-protocol, but
the ocaml bindings don't generate optional arguments, so we need
to pass this parameter every time we call login_with_password.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>