CP-15214: Windows: device 0002 not 0001 after all.
This reverses the spirit of CP-14788: it turns out that the Windows
templates still need the 0002 PCI device rather than 0001.

According to Paul Durrant,
"Apparently PVS expects that it is possible to take a disk image
from a VM provisioned on one version of XenServer and use it on a VM
running on another version of XenServer. This basically means that
we can never modify the virtual hardware platform, only add to it."

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>